# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI SkillFit API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get recent interview activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
...
    ActivityItem:
      type: object
      properties:
        type:
          type: string
          enum: [interview_completed, candidate_registered, action_taken]
        candidateId:
          type: integer
        candidateName:
          type: string
        trade:
          type: string
        classification:
          type: string
          nullable: true
        action:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - type
        - candidateId
        - candidateName
        - trade
        - timestamp